Fixed the chronometer options
The chronometer options where send, but not in the correct format. In the documentation it was also not clear how and what. So now when sending ```data_chronometer: "{{now()+timedelta(minutes=30)}}"``` the value is converted to a timestamp and then converted to an integer and added to the data as an string 🥴