Android Auto features (#4)
* Android Auto features

This version adds two new options for Android Auto (Android Devices only):

1. Added the `car_ui: true` when the device is Android and when the android_auto binary sensor is on, so you can interact with the notifications whilst driving.
https://companion.home-assistant.io/docs/notifications/notifications-basic/#android-auto-visibility-beta
2. Added a new option: `Android Only - Notify using TTS whilst using Andriod Auto? (Optional)`
This "upgrades" the notification message to a TTS message, when the `binary_sensor.###_android_auto` is `on`. This can be useful whilst driving with Android Auto that the notification that you are sending is announced via TTS.

* Fixed a typo

* Added a new section for Android Auto

The blueprint now contains a input `Android Only - Enter a custom Text to Speech text used to announce the actions, when your device is connected to the Android Auto.` . 

This is only relevant if all of the following are true: 
* Your using a Android phone
* You enabled the new option `Android Only - Notify using TTS whilst using Android Auto? (Optional)`
* Your Android phone is in the Android Auto mode (when the `binary_sensor.###_android_auto` is `on`)
* When you have actions defined, that could require action from the user

This new feature allows you to hear the Text to Speech message, receive the notification and the possible interact with actions that you could take. The input field is useful to customize the TTS message in the desired language. The format to use is as follows:
`message`+`announcement`+`actions`
Only the announcement part is needed as the input to introduce the actions.
Examples:
* Select 1 of the following actions on your phone
* Selecteer 1 van de volgende acties op je telefoon


## Use case example ##
```
service: script.notify_devices
data:
  notify_message: The alarm is turned off
  data_tag: alarm
  data_channel: Alarm
  data_group: Alarm
  data_sticky: true
  data_importance: high
  data_timeout: 300
  notify_title: Alarm • Uit
  data_notification_icon: mdi:shield-lock-open-outline
  notify_home_or_away: Both
  data_clickaction_url: /dashboard-alarm/alarm
  data_notification_color:
    - 0
    - 173
    - 0
  data_persistent: true
  notify_in_car_using_tts: true
  data_actions: >-
    {{iif(is_state('binary_sensor.alarm_night_mode','off'),[{'action':'RESUME_NIGHT_ALARM','title':'Enable the night alarm'},],'')}}

```
This would result in the following TTS message:
`The alarm is turned off. Select 1 of the following actions on your phone: Enable the night alarm`