Improved devices validation
This change will make sure that only correct devices, registered via the Mobile_App, will be notified. When passing any device, then in the previous version it would have tried to send a notification to it, but when the device is not part of the Mobile_App integration then it would fail. This fix fixes that issue, so only correct device_trackers can be used.